+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+- Move the app config management APIs to the /admin subrouter from /system [#652](https://github.com/rokwire/core-building-block/issues/652)
 ### Added
 - Enable CORS [#632](https://github.com/rokwire/core-building-block/issues/632)